An initial multiplayer test. This will allow multiple "clients" to join the game at one time, and draw them all together.
This is primarly meant as a lag-test. If this "plays" well over the LAN, we will use this as the foundation for our multiplayer support.

If you would like to connect to the server, you'll need to either (a) run your own node server based on server/server.js (b) edit app/Multiplayer.js to connect to the correct server.