Working island
Implemented a basic mask that removes terrain farther from the enter of
terrain, giving the appearance of an island. It still needs a lot of
work.

Lastly, doubled the number of points to make the island smoother

I also added camera controls to move up and down so I could inspect the
terrain. This needs t be removed someday.