Tweaked gameplay
It's now easier to be more precise. The turret turns more slowly,
projectiles are heavier, and projectiles have a lower initial velocity.

Though this does deviate from our original idea for the game, the tanks
and projectiles are so much smaller than the map that sticking with the
original settings seems like a bad idea.